requirements.txt
six>=1.10.0
splunk-sdk>=1.7.2
python-dotenv==0.19.2
requests>=2.29.0
urllib3~=2.2.3
loguru~=0.7.2